node-back/requirements.txt
joblib==1.5.2
mysql-connector-python==9.5.0
numpy==2.3.4
pandas==2.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2

